+++ src/constants/index.ts
@@ -4,10 +4,6 @@ export const MAX_GAS_LIMIT = '600000000';
 export const GAS_LIMIT_DELTA = 10;
 export const TOKEN_GAS_LIMIT = '500000';
 export const MIN_DUST = '5000000000000000'; // 0.05 EGLD
-export const LEDGER_MULTI_ACCOUNT_MINIMUM_VERSION = '1.0.8';
-export const LEDGER_HASH_SIGN_MINIMUM_VERSION = '1.0.11';
-export const LEDGER_SIGN_AUTH_TOKEN_MINIMUM_VERSION = '1.0.15';
-export const LEDGER_WITH_WHITELISTED_TOKENS_MINIMUM_VERSION = '1.0.17';
 
 export const ZERO = '0';
 export const WEGLD_ID = 'WEGLD-bd4d79';