+++ src/operations/index.ts
@@ -9,6 +9,5 @@ export * from './computeDataField';
 export * from './computeInitGasLimit';
 export * from './formattedConfigGasPrice';
 export * from './getEntireBalance';
-export * from './ledger';
 export * from './getGasLimit';
 export * from './getGuardedAccountGasLimit';