+++ src/validationSchema/data.ts
@@ -1,6 +1,6 @@
+import { LEDGER_HASH_SIGN_MINIMUM_VERSION } from '@multiversx/sdk-dapp/constants/ledger.constants';
+import getLedgerVersionOptions from '@multiversx/sdk-dapp/utils/operations/ledger/getLedgerVersionOptions';
 import { string } from 'yup';
-import { LEDGER_HASH_SIGN_MINIMUM_VERSION } from 'constants/index';
-import { getLedgerVersionOptions } from 'operations';
 import { ExtendedValuesType } from 'types';
 
 const ledgerDataActive = string().test(
@@ -20,13 +20,20 @@ const ledgerDataActive = string().test(
 const hashSign = string().test({
   name: 'hashSign',
   test: function hashSignCheck(value) {
-    const { ledger } = this.parent as ExtendedValuesType;
+    const { ledger, isGuarded } = this.parent as ExtendedValuesType;
 
     if (ledger) {
-      const { ledgerWithHashSign } = getLedgerVersionOptions(ledger.version);
+      const { ledgerWithHashSign, ledgerWithGuardians } =
+        getLedgerVersionOptions(ledger.version);
       if (value && value.length > 300 && !ledgerWithHashSign) {
         return this.createError({
           message: `Data too long. You need at least MultiversX app version ${LEDGER_HASH_SIGN_MINIMUM_VERSION}. Update MultiversX app to continue`,
+          path: 'data'
+        });
+      }
+      if (isGuarded && !ledgerWithGuardians) {
+        return this.createError({
+          message: `You need at least MultiversX app version ${LEDGER_HASH_SIGN_MINIMUM_VERSION} to use Guardians`,
           path: 'data'
         });
       }