+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- [Verify ledger version for guarded transactions](https://github.com/multiversx/mx-sdk-dapp-form/pull/171)
 - [Extend guarded account gasLimit validation for ESDT and NFT](https://github.com/multiversx/mx-sdk-dapp-form/pull/170)
+
 
 ## [[0.5.3](https://github.com/multiversx/mx-sdk-dapp-form/pull/169)] - 2023-05-03
 - [Fixed setting Guardian screen for prefilled transactions](https://github.com/multiversx/mx-sdk-dapp-form/pull/168)